--- src/hooks/api/retrospective/useCreateAnswerMutation.ts
@@ -1,10 +1,0 @@
-import { useMutation } from '@tanstack/react-query';
-
-import { createRetrospectiveAnswer } from '@/apis/pull-requests/retrospective.mutate';
-
-export const useCreateAnswerMutation = () =>
-  useMutation({
-    mutationFn: ({ questionId }: { questionId: number }) => {
-      return createRetrospectiveAnswer(questionId);
-    },
-  });
